requirements.txt
openai
google-genai
python-dotenv
pypdf
pandas
openpyxl
PyYAML
pytest

